DDF-1366 Moved the SourcePoller from core-api-impl to standard-framework
