+++ catalog/core/catalog-core-api/src/main/java/ddf/catalog/source/Source.java
@@ -37,7 +37,6 @@ public interface Source extends Describable {
    * periodically.
    *
    * @return true - if the site is available (up), false - if the site is unavailable (down)
-   * @see ddf.catalog.util.impl.SourcePoller
    */
   public boolean isAvailable();
