+++ catalog/core/catalog-core-standardframework/src/main/java/org/codice/ddf/catalog/sourcepoller/Poller.java
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package ddf.catalog.util.impl;
+package org.codice.ddf.catalog.sourcepoller;
 
 import static org.apache.commons.lang3.Validate.notNull;
