+++ catalog/core/catalog-core-standardframework/src/test/groovy/org/codice/ddf/catalog/sourcepoller/PollerRunnerSpec.groovy
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package ddf.catalog.util.impl
+package org.codice.ddf.catalog.sourcepoller
 
 import com.google.common.collect.ImmutableMap
 import org.spockframework.runtime.SpockAssertionError
