+++ catalog/core/catalog-core-standardframework/src/test/groovy/org/codice/ddf/catalog/sourcepoller/StatusSourcePollerSpec.groovy
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package ddf.catalog.util.impl
+package org.codice.ddf.catalog.sourcepoller
 
 import spock.lang.Specification
