+++ catalog/core/catalog-core-standardframework/src/test/java/ddf/catalog/impl/CatalogFrameworkQueryTest.java
@@ -52,8 +52,6 @@ import ddf.catalog.source.IngestException;
 import ddf.catalog.source.Source;
 import ddf.catalog.source.SourceUnavailableException;
 import ddf.catalog.source.UnsupportedQueryException;
-import ddf.catalog.util.impl.SourcePoller;
-import ddf.catalog.util.impl.SourceStatus;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Collections;
@@ -62,6 +60,8 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import org.codice.ddf.catalog.sourcepoller.SourcePoller;
+import org.codice.ddf.catalog.sourcepoller.SourceStatus;
 import org.codice.ddf.platform.util.uuidgenerator.UuidGenerator;
 import org.geotools.filter.FilterFactoryImpl;
 import org.geotools.temporal.object.DefaultInstant;
