+++ distribution/docs/src/main/resources/content/_reference/_tables/sourcePoller.adoc
@@ -1,11 +1,11 @@
 :title: Status Source Poller Runner
-:id: ddf.catalog.util.impl.StatusSourcePollerRunner
+:id: org.codice.ddf.catalog.sourcepoller.StatusSourcePollerRunner
 :status: published
 :type: table
 :application: ${ddf-catalog}
 :summary: Status Source Poller Runner.
 
-.[[_ddf.catalog.util.impl.StatusSourcePollerRunner]]Status Source Poller Runner
+.[[_org.codice.ddf.catalog.sourcepoller.StatusSourcePollerRunner]]Status Source Poller Runner
 [cols="1,1m,1,3,1,1" options="header"]
 |===
